Some mISDN fixes.
* Fixed T305 timer not started when another state's timeout transitions to
the U11 state.

* Made l3dss1_setup_req() always transition to state U1.  We don't care if
sending the SETUP message failed.  We need to be in state U1 to clean up
properly.

* Removes state change code from SendMsg().  If SendMsg() failed you don't
know if the state was changed.  It also should never have been in that
function in the first place.

* Fixed type mismatch in state comparison for valid HOLD states.

* Fixed call ref selection to try to avoid assigning a call ref that is
already in use.  Also removed sending the RELEASE COMPLETE when a SETUP
does not get any response because the call ref may already be in use by
another call.

* Fixed a l3m_debug() message erroneously truncating the call ref.

* Some indentation fixes.

The jira_abe_2869_misdn_fixes.patch started with the t305.patch file from
AST-914.

(issue ABE-2869)
Patches:
      jira_abe_2869_misdn_fixes.patch (license #5621) patch uploaded by rmudgett