Fix mISDN B channels getting lost over time.
Sometimes messages sent by Asterisk in isdn_lib.c to the kernel driver
never reach it.  In case of these stuck B channels it is always first a
SETUP which is never acted upon (no TIMEOUT/no SETUP-ACK indication
follows) and then later the DISCONNECT is also not answered with an
appropriate RELEASE.

The problem was introduced in mISDN on 2004-11-19 with a "fix" for kernel
2.6.8, which turns out to be plain wrong.  Before kernel 2.6.8 it was
required to check "if (off != &file->f_pos)" and return -ESPIPE if a
device is not seekable.  Since 2.6.8, this check does not work anymore.
The fix for kernel 2.6.8 is to use the new API function nonseekable_open()
to disallow seeks.

(closes issue ABE-2869)
Reported by: Guenther Kelleter
Patches:
      noseek.diff (license #6372) patch uploaded by Guenther Kelleter